Agrego interfaz y nuevas clases

+++ src/ar/edu/unq/po2/tp5/ProductoCooperativa.java
@@ -2,13 +2,13 @@ package ar.edu.unq.po2.tp5;
 
 public class ProductoCooperativa  extends Producto{
 
-	public ProductoCooperativa(String nombre, double precio, int stock) {
+	public ProductoCooperativa(String nombre, float precio, int stock) {
 		super(nombre, precio, stock);
 	}
 	
 	@Override
-	public double getPrecio() {
-		return super.getPrecio() * 0.9;
+	public float getPrecio() {
+		return (float) (super.getPrecio() * 0.9);
 	}
  
 }
